+++ R/IntCal.R
@@ -4,7 +4,7 @@
 #' This data package provides the files of these curves, for use by other R package (reducing the need for replication and the size of other packages that use IntCal).
 #' It also comes with a limited number of relevant functions, to read in calibration curves, translate pMC ages to 14C ages (et vice versa), etc. 
 #' @docType package
-#' @author Maarten Blaauw <maarten.blaauw@qub.ac.uk> 
+#' @author Maarten Blaauw <maarten.blaauw@qub.ac.uk>
 #' @importFrom utils read.csv read.table write.table packageName
 #' @importFrom stats approx dnorm
 #' @importFrom grDevices rgb extendrange
@@ -12,7 +12,7 @@
 #' @name IntCal
 NULL
 
-# todo: the cal.dist polygon of calibrate(-900,90, cc="nh1") doesn't look well. Make calibrate (or ccurve?) work better with postbomb curve 
+# todo: 
 
 
 #' @name list.ccurves
@@ -149,136 +149,33 @@ ccurve <- function(cc=1, postbomb=FALSE) {
 
 
 
-#' @name draw.ccurve
-#' @title Draw a calibration curve.
-#' @description Draw one or two of the calibration curves, or add a calibration curve to an existing plot.
-#' @return The calibration curve (invisible).
-#' @param cal1 First calendar year for the plot
-#' @param cal2 Last calendar year for the plot
-#' @param cc1 Name of the calibration curve. Can be "IntCal20", "Marine20", "SHCal20", or for the previous curves "IntCal13", "Marine13" or "SHCal13". Can also be "nh1", "nh2", "nh3", "sh1-2", "sh3", "Kure", "LevinKromer" or "Santos" for postbomb curves.
-#' @param cc2 Optional second calibration curve to plot. Can be "IntCal20", "Marine20", "SHCal20", or for the previous curves "IntCal13", "Marine13" or "SHCal13". Defaults to nothing, NA.
-#' @param cc1.postbomb Use \code{postbomb=TRUE} to get a postbomb calibration curve for cc1 (default \code{cc1.postbomb=FALSE}).
-#' @param cc2.postbomb Use \code{postbomb=TRUE} to get a postbomb calibration curve for cc2 (default \code{cc2.postbomb=FALSE}).
-#' @param BCAD The calendar scale of graphs and age output-files is in cal BP (calendar or calibrated years before the present, where the present is AD 1950) by default, but can be changed to BC/AD using \code{BCAD=TRUE}.
-#' @param cal.lab The labels for the calendar axis (default \code{age.lab="cal BP"} or \code{"BC/AD"} if \code{BCAD=TRUE}), or to \code{age.lab="kcal BP"} etc. if ka=TRUE.
-#' @param cal.rev Reverse the calendar axis. 
-#' @param c14.lab Label for the C-14 axis. Defaults to 14C BP (or 14C kBP if ka=TRUE).
-#' @param c14.lim Axis limits for the C-14 axis. Calculated automatically by default. 
-#' @param c14.rev Reverse the C-14 axis.
-#' @param ka Use kcal BP (and C14 kBP).
-#' @param add.yaxis Whether or not to plot the second calibration. Defaults to \code{add.yaxis=FALSE}.
-#' @param cc1.col Colour of the calibration curve (outline).
-#' @param cc1.fill Colour of the calibration curve (fill).
-#' @param cc2.col Colour of the calibration curve (outline), if activated (default cc2=NA).
-#' @param cc2.fill Colour of the calibration curve (fill), if activated (default cc2=NA).
-#' @param add Whether or not to add the curve(s) to an existing plot. Defaults to FALSE, which draws a new plot
-#' @param bty Draw a box around a box of a certain shape. Defaults to \code{bty="l"}.
-#' @param ... Any additional optional plotting parameters. 
-#' @examples 
-#' draw.ccurve()
-#' draw.ccurve(1000, 3000, cc2="Marine20")
-#' draw.ccurve(1800, 2020, BCAD=TRUE, cc2="nh1", cc2.postbomb=TRUE)
-#' draw.ccurve(1800, 2010, BCAD=TRUE, cc2="nh1", add.yaxis=TRUE)
-#' @export
-draw.ccurve <- function(cal1=-50, cal2=55e3, cc1="IntCal20", cc2=NA, cc1.postbomb=FALSE, cc2.postbomb=FALSE,  BCAD=FALSE, cal.lab=NA, cal.rev=FALSE, c14.lab=NA, c14.lim=NA, c14.rev=FALSE, ka=FALSE, add.yaxis=FALSE, cc1.col=rgb(0,0,1,.5), cc1.fill=rgb(0,0,1,.2), cc2.col=rgb(0,.5,0,.5), cc2.fill=rgb(0,.5,0,.2), add=FALSE, bty="l", ...) {
-
-  # read and narrow down the calibration curve(s)
-  cc.1 <- ccurve(cc1, cc1.postbomb)
-  if(BCAD)
-    cc.1[,1] <- 1950 - cc.1[,1] 
-  mindat <- cc.1[,1] >= min(cal1, cal2)
-  maxdat <- cc.1[,1] <= max(cal1, cal2)
-  cc.1 <- cc.1[which(mindat * maxdat == 1),]
-  if(ka)
-    cc.1 <- cc.1/1e3
-  cc1.pol <- cbind(c(cc.1[,1], rev(cc.1[,1])), c(cc.1[,2]-cc.1[,3], rev(cc.1[,2]+cc.1[,3])))
-  
-  if(!is.na(cc2)) {
-    cc.2 <- ccurve(cc2, cc2.postbomb)
-    if(BCAD)
-      cc.2[,1] <- 1950 - cc.2[,1] 
-    mindat <- cc.2[,1] >= min(cal1, cal2)
-    maxdat <- cc.2[,1] <= max(cal1, cal2)
-    cc.2 <- cc.2[which(mindat * maxdat == 1),]
-    if(ka)
-      cc.2 <- cc.2/1e3
-    cc2.pol <- cbind(c(cc.2[,1], rev(cc.2[,1])), c(cc.2[,2]-cc.2[,3], rev(cc.2[,2]+cc.2[,3])))
-  }
-
-  if(!add) { # then prepare plotting parameters
-    if(is.na(cal.lab))
-      callab <- "cal. yr BP"
-    if(is.na(c14.lab))
-      c14lab <- expression(""^14*C~BP)
-
-    cal.lim <- c(cal1, cal2)
-    if(cal.rev)
-      cal.lim <- rev(cal.lim)
-
-    if(BCAD)
-      if(is.na(cal.lab))
-        callab <- "BC/AD"
-
-    if(ka) {
-      if(is.na(c14.lab))
-        c14lab <- expression(""^14*C~kBP)
-      if(is.na(cal.lab))
-        callab <- ifelse(BCAD, "kcal BC/AD", "kcal BP")
-      cal.lim <- cal.lim/1e3
-    }
-
-    if(is.na(c14.lim))
-      if(is.na(cc2) || add.yaxis)
-        c14.lim <- range(cc1.pol[,2]) else
-          c14.lim <- range(cc1.pol[,2], cc2.pol[,2])
-    if(c14.rev)
-      c14.lim <- rev(c14.lim)
-
-    # draw the graph and data
-    plot(0, type="n", xlim=cal.lim, xlab=callab, ylim=c14.lim, ylab=c14lab, bty=bty, ...)
-  }
-
-  # add the calibration curve
-  polygon(cc1.pol, col=cc1.fill, border=NA) # calibration curve
-  lines(cc.1[,1], cc.1[,2]-cc.1[,3], col=cc1.col)
-  lines(cc.1[,1], cc.1[,2]+cc.1[,3], col=cc1.col)
-
-  # add a second curve?
-  if(!is.na(cc2)) {
-    if(add.yaxis) {
-      op <- par(new=TRUE)
-      plot(cc2.pol, type="n", xlim=cal.lim, xlab="", ylab="", bty="n", xaxt="n", yaxt="n")
-    }
-    polygon(cc2.pol, col=cc2.fill, border=NA) # calibration curve
-    lines(cc.2[,1], cc.2[,2]-cc.2[,3], col=cc2.col)
-    lines(cc.2[,1], cc.2[,2]+cc.2[,3], col=cc2.col)
-    if(add.yaxis)
-      axis(4, col=cc2.col, col.axis=cc2.col)
-  }
-}
-
-
-
-
 #' @name mix.curves
 #' @title Build a custom-made, mixed calibration curve.
-#' @description If two curves need to be `mixed' to calibrate, e.g. for dates of mixed terrestrial and marine carbon sources, then this function can be used. The curve will be saved together with the calibration curves. 
+#' @description If two curves need to be `mixed' to calibrate, e.g. for dates of mixed terrestrial and marine carbon sources, then this function can be used. The curve will be saved, together with the main calibration curves, in a temporary directory. This temporary directory then has to be specified in further commands, e.g. for rbacon: \code{Bacon(, ccdir=tmpdr)} (see examples). It is advisable to make your own curves folder and have ccdir point to that folder. 
 #' @details The proportional contribution of each of both calibration curves has to be set.
 #'
 #' @param proportion Proportion of the first calibration curve required. e.g., change to \code{proportion=0.7} if \code{cc1} should contribute 70\% (and \code{cc2} 30\%) to the mixed curve.
 #' @param cc1 The first calibration curve to be mixed. Defaults to the northern hemisphere terrestrial curve IntCal20.
 #' @param cc2 The second calibration curve to be mixed. Defaults to the marine curve IntCal20.
 #' @param name Name of the new calibration curve.
-#' @param dir Optional name of the directory where to save the file. Defaults to where the IntCal curves live.
+#' @param dir Name of the directory where to save the file. Since R does not allow automatic saving of files, this points to a temporary directory by default. Adapt to your own folder, e.g., \code{dir="~/Curves"} or in your current working directory, \code{dir="."}.
 #' @param offset Any offset and error to be applied to \code{cc2} (default 0 +- 0).
 #' @param sep Separator between fields (tab by default, "\\t")
 #' @return A file containing the custom-made calibration curve, based on calibration curves \code{cc1} and \code{cc2}.
 #' @examples
 #' mix.curves()
+#' tmpdir <- tempdir()
+#' tmpdir
+#' mix.curves(dir=tmpdir)
+#' # clean up:
+#' file.remove(tmpdir)
 #' @export
 mix.curves <- function(proportion=.5, cc1="IntCal20", cc2="Marine20", name="mixed.14C", dir=c(), offset=c(0,0), sep="\t") {
+  # place the IntCal curves within the same folder as the new curve:
   if(length(dir) == 0)
-    dir <- system.file("extdata/", package='IntCal')
+     dir <- tempdir()   
+  curves <- list.files(system.file("extdata", package='IntCal'), pattern=".14C", full.names=T)
+  file.copy(curves, dir)
 
   cc1 <- ccurve(cc1)
   cc2 <- ccurve(cc2)
@@ -287,8 +184,7 @@ mix.curves <- function(proportion=.5, cc1="IntCal20", cc2="Marine20", name="mixe
   cc2.error <- sqrt(cc2.error^2 + offset[2]^2)
   mu <- proportion * cc1[,2] + (1-proportion) * cc2.mu
   error <- proportion * cc1[,3] + (1-proportion) * cc2.error
+  
   write.table(cbind(cc1[,1], mu, error), paste0(dir, name), row.names=FALSE, col.names=FALSE, sep=sep)
+  message(name, " saved in folder ", dir)
 }
-
-
-